Challenge: shuffleCards (Solution)
Here is my solution for the shuffleCards recursion challenge. Some challenges I faced during the solving process was tracking what element of the top of bottom half of the deck to push into the shuffled array. After some research I was able to propery control the flow of pushing and what to increment instead of mutating the original arrays.